refactor(iife): extract separable logic
